+++ CHANGES.md
@@ -6,7 +6,12 @@
   Thanks to Thomas Gazagnaire for report (#13).
 
 - Add `Bytes.Writer.writes_to_string`
+
 - `Bytesrw_sysrandom`: fix headers for musl libc (alpine.) (#7)
+
+- Deprecate `Bytes.Slice.{take,drop,break}` in favor of 
+  `Bytes.Slice.{take_first,drop_first,cut_first}` to align on the stdlib's 
+  new `String` terminology (#11).
 
 v0.3.0 2025-11-04 Zagreb
 ------------------------